requirements.txt
obs-websocket-py==0.5.1
